pass screen coordinates to shader

+++ source/os/include/os_graphics.h
@@ -7,7 +7,7 @@ void os_graphics_create_texture(OSWindowHandle os_window_handle, const u32* text
 void os_graphics_create_vertex_shader(OSWindowHandle os_window_handle, const u8* shader_buffer, u32 shader_buffer_size);
 void os_graphics_create_pixel_shader(OSWindowHandle os_window_handle, const u8* shader_buffer, u32 shader_buffer_size);
 void os_graphics_clear(OSWindowHandle os_window_handle, f32 r, f32 g, f32 b, f32 a);
-void os_graphics_draw_rectangle(OSWindowHandle os_window_handle, f32 x, f32 y, f32 width, f32 height, f32 r, f32 g, f32 b);
+void os_graphics_draw_rectangle(OSWindowHandle os_window_handle, i32 x, i32 y, i32 width, i32 height, u8 r, u8 g, u8 b);
 void os_graphics_draw(OSWindowHandle os_window_handle);
 
 #define H_OS_GRAPHICS_H
